fixed file reading added routes
