+++ src/main.rs
@@ -1,5 +1,5 @@
-mod brick;
-
+mod robot;
 fn main() {
-    println!("{}", brick::file_read_int("/sys/class/tacho-motor/motor0/position"));
+    
+    
 }
