+++ src/robot/brick.rs
@@ -1,0 +1,52 @@
+use std::io::Error;
+use std::fs::File;
+use std::fs::OpenOptions;
+use std::io::{BufReader, Read, Write};
+
+// functions for handeling file reads
+pub fn file_read_string(path: &str) -> String {
+    let file_result: Result<File, Error> = File::open(path);
+    let file:File = match file_result {
+        Ok(file) => file,
+        Err(error) => panic!("Faild to read file: {:?}, Error: {:?}", path, error)
+    };
+    let mut file_buff:BufReader<File> = BufReader::new(file);
+
+    let mut data:String = String::new();
+    file_buff.read_to_string(&mut data).expect("Faild to read string");
+    return data;
+}
+
+pub fn file_read_int(path: &str) -> i64 {
+    let file_result: Result<File, Error> = File::open(path);
+    let file:File = match file_result {
+        Ok(file) => file,
+        Err(error) => panic!("Faild to read file: {:?}, Error: {:?}", path, error)
+    };
+    let mut file_buff = BufReader::new(file);
+
+    let mut raw_data:String = String::new();
+    file_buff.read_to_string(&mut raw_data).expect("Faild to read int");
+
+    let data: i64 = raw_data.trim().parse::<i64>().unwrap(); // not to self: do not forget to trim the string when type casting
+    return data;
+}
+
+// function for handeling file writes
+pub fn file_write_string(path: &str, data: &str) {
+    let file_result:Result<File, Error> = OpenOptions::new().write(true).open(path);
+    let mut file:File = match file_result {
+        Ok(file) => file,
+        Err(error) => panic!("Faild to write file: {:?} data: {:?} Error: {:?}", path, data, error)
+    };
+    file.write_all(format!("{}", data).as_bytes()).unwrap();
+}
+
+pub fn file_write_int(path: &str, data: i64) {
+    let file_result:Result<File, Error> = OpenOptions::new().write(true).open(path);
+    let mut file:File = match file_result {
+        Ok(file) => file,
+        Err(error) => panic!("Faild to write file: {:?} data: {:?} Error: {:?}", path, data, error)
+    };
+    file.write_all(format!("{}", data).as_bytes()).unwrap();
+}
